Suggested Changes Done. TODO: Env Format Change
